fix(mod): capture correct time-of-day ticks, add human-readable time/moon formatting
SceneExporter was sending raw dayTime() (total world age) instead of
dayTime % 24000. Narrow PackageEnvironment.time to Int, add shared
SceneFormatting utilities, and replace all inline time/moon display
across mod UI and frontend with consistent clock + moon phase names.

+++ mod/common/src/main/kotlin/com/xevion/glint/api/SceneSyncManager.kt
@@ -228,7 +228,7 @@ object SceneSyncManager {
                                 ),
                             environment =
                                 UploadEnvironment(
-                                    timeOfDayTicks = metadata.environment.time.toInt(),
+                                    timeOfDayTicks = metadata.environment.time,
                                     weather = metadata.environment.weather,
                                     weatherIntensity = metadata.environment.weatherIntensity.toDouble(),
                                     moonPhase = metadata.environment.moonPhase,
